[refactor] #30 query service 파일 위치 및 변수명 수정
